API: refactor to avoid modifying auth.py
I didn't lie the idea of the user modifying auth.py
to implement get_password_hash_for_username, and also, it
didn't make much sense messing with db stuff here. So I've
moved the function to models_db.py so that it lives with the
rest of the persistence layer functionality.